Add "Upspring -run whatever.lua", fix a lot related stuff

+++ src/upspring/ObjectView.cpp
@@ -67,8 +67,8 @@ fltk::Widget *EditorUI::ObjectView::List::child (const fltk::Menu *, const int *
 	MdlObject *obj = root_obj();
 	if (!obj) return 0;
 	for (;;) {
-		int i = *indexes++;
-		if (i < 0 || i >= int(obj->childs.size())) return 0;
+		size_t i = *indexes++;
+		if (i < 0 || i >= obj->childs.size()) return 0;
 		MdlObject *ch = obj->childs[i];// Widget* widget = group->child(i);
 		if (!level--) {
 			obj = ch;
